refactor(mobile): split MobileApp god file and unify surface state
