+++ packages/ui/src/apps/MobileOverflowMenu.tsx
@@ -1,0 +1,79 @@
+import React from 'react';
+
+import { Icon } from '@/components/icon/Icon';
+import type { IconName } from '@/components/icon/icons';
+import { useI18n } from '@/lib/i18n';
+import { cn } from '@/lib/utils';
+
+export type OverflowItem = {
+  key: 'files' | 'changes' | 'terminal' | 'mcp' | 'instances' | 'update' | 'settings';
+  icon?: IconName;
+  iconNode?: React.ReactNode;
+  label: string;
+  badge?: number;
+  onSelect: () => void;
+};
+
+export const MobileOverflowMenu: React.FC<{
+  open: boolean;
+  onClose: () => void;
+  items: OverflowItem[];
+  /** Extra viewport-right inset so the dropdown stays anchored to the
+      three-dots button when the iPad right sidebar shifts the header. */
+  rightOffset?: number;
+}> = ({ open, onClose, items, rightOffset = 0 }) => {
+  const { t } = useI18n();
+  React.useEffect(() => {
+    if (!open) return;
+    const handleKey = (event: KeyboardEvent) => {
+      if (event.key === 'Escape') onClose();
+    };
+    document.addEventListener('keydown', handleKey);
+    return () => document.removeEventListener('keydown', handleKey);
+  }, [onClose, open]);
+
+  if (!open) return null;
+
+  return (
+    <div className="fixed inset-0 z-50" role="dialog" aria-modal="true" aria-label={t('mobile.menu.titleAria')}>
+      <button
+        type="button"
+        className="absolute inset-0 cursor-default"
+        aria-label={t('mobile.surface.closeAria')}
+        onClick={onClose}
+      />
+      <div
+        className="absolute top-[calc(var(--oc-safe-area-top,0px)+56px+4px)] w-[min(220px,calc(100vw-1rem))] origin-top-right overflow-hidden rounded-2xl border border-border/40 bg-background shadow-[0_18px_60px_rgb(0_0_0_/_0.35)]"
+        role="menu"
+        style={{
+          right: `${8 + rightOffset}px`,
+          animation: 'mobile-menu-in 160ms cubic-bezier(0.32, 0.72, 0, 1)',
+        }}
+      >
+        {items.map((item, index) => (
+          <button
+            key={item.key}
+            type="button"
+            role="menuitem"
+            className={cn(
+              'flex w-full items-center gap-3 px-4 py-3 text-left transition-colors hover:bg-interactive-hover focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:ring-inset',
+              index > 0 && 'border-t border-border/30',
+            )}
+            style={{ touchAction: 'manipulation' }}
+            onClick={() => {
+              item.onSelect();
+              onClose();
+            }}
+          >
+            {item.iconNode ?? (item.icon ? <Icon name={item.icon} className="size-5 shrink-0 text-muted-foreground" /> : null)}
+            <span className="min-w-0 flex-1 truncate typography-ui-label text-foreground">{item.label}</span>
+            {item.badge && item.badge > 0 ? (
+              <span className="inline-flex size-2 shrink-0 rounded-full bg-primary" aria-hidden />
+            ) : null}
+          </button>
+        ))}
+      </div>
+      <style>{`@keyframes mobile-menu-in { from { opacity: 0; transform: translateY(-6px) scale(0.96); } to { opacity: 1; transform: translateY(0) scale(1); } }`}</style>
+    </div>
+  );
+};
